Extract syscall_fstat() handler
